requirements.txt
Werkzeug==1.0.0
tensorflow==1.14.0
requests==2.22.0
opencv_python==4.5.1.48
imageio==2.6.1
numpy==1.19.2
scipy==1.6.2
six==1.15.0
torch==1.8.1
h5py==2.10.0
torchvision==0.9.1
tqdm==4.42.1
dlib==19.22.0
tensorboard==1.14.0
flask_restplus==0.13.0
Flask==1.1.1
apex==0.9.10dev
dominate==2.6.0
Image==1.5.33
lmdb==1.2.1
Pillow==8.2.0
pycuda==2021.1
scikit_learn==0.24.1
tensorrt==0.0.1
